modified:   app/Http/Controllers/BSITStudentsController.php 	new file:   app/Http/Controllers/MainController.php 	new file:   app/Models/Report.php 	modified:   database/migrations/2025_03_07_172416_create_students_table.php 	new file:   database/migrations/2025_04_14_052409_create_reports_table.php 	deleted:    resources/views/blis.blade.php 	deleted:    resources/views/bscs.blade.php 	deleted:    resources/views/create.blade.php 	deleted:    resources/views/create_blis.blade.php 	deleted:    resources/views/create_bscs.blade.php 	new file:   resources/views/edit.blade copy.php 	modified:   resources/views/edit.blade.php 	modified:   resources/views/layouts/app.blade.php 	modified:   resources/views/layouts/navigation.blade.php 	new file:   resources/views/reports.blade.php 	renamed:    resources/views/bsit.blade.php -> resources/views/show.blade.php 	modified:   resources/views/students/create.blade.php 	modified:   routes/web.php

+++ database/migrations/2025_03_07_172416_create_students_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('fullname');
             $table->string('address');
-            $table->integer('contact_number');
+            $table->bigInteger('contact_number');
             $table->string('school');
             $table->string('strand');
             $table->integer('age');
